roo/kafka_learning_project/requirements.txt
confluent-kafka-python